+++ apps/condo/index.js
@@ -19,7 +19,8 @@ const {
 } = require('@open-condo/keystone/healthCheck')
 const { prepareKeystone } = require('@open-condo/keystone/KSv5v6/v5/prepareKeystone')
 const { RequestCache } = require('@open-condo/keystone/requestCache')
-const { getWebhookModels, getWebhookTasks } = require('@open-condo/webhooks/schema')
+const { getWebhookModels } = require('@open-condo/webhooks/schema')
+const { getWebhookTasks } = require('@open-condo/webhooks/tasks')
 
 const { PaymentLinkMiddleware } = require('@condo/domains/acquiring/PaymentLinkMiddleware')
 const FileAdapter = require('@condo/domains/common/utils/fileAdapter')
