+++ packages/webhooks/plugins/webHooked.js
@@ -1,7 +1,7 @@
 const { composeNonResolveInputHook } = require('@open-condo/keystone/plugins/utils')
 const { GQL_SCHEMA_PLUGIN } = require('@open-condo/keystone/plugins/utils/typing')
 const { getModelValidator } = require('@open-condo/webhooks/model-validator')
-const { getWebhookTasks } = require('@open-condo/webhooks/schema')
+const { getWebhookTasks } = require('@open-condo/webhooks/tasks')
 
 const plugin = (fn) => {
     fn._type = GQL_SCHEMA_PLUGIN
