+++ packages/webhooks/tasks/sendModelWebhooks.js
@@ -1,7 +1,7 @@
 const { find } = require('@open-condo/keystone/schema')
 
 async function sendModelWebhooks (modelName) {
-    const { getWebhookTasks } = require('@open-condo/webhooks/schema')
+    const { getWebhookTasks } = require('@open-condo/webhooks/tasks')
     // TODO(DOMA-4570): Add filter for only responsive subscriptions (failCount < threshold)
     // NOTE: find to bypass default limitations, since relations are not used here
     const subscriptions = await find('WebhookSubscription', {
